requirements.txt
aniso8601==7.0.0
annotated-types==0.6.0
anyio==3.7.1
asyncpg==0.28.0
blinker==1.6.3
click==8.1.7
dateparser==1.1.8
dnspython==2.4.2
docutils==0.20.1
fastapi==0.103.2
Flask==3.0.0
graphene==2.1.9
graphene-sqlalchemy==2.3.0
graphql-core==3.2.3
graphql-relay==2.0.1
greenlet==3.0.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lockfile==0.12.2
Markdown==3.5
MarkupSafe==2.1.3
mock==5.1.0
note==0.5.2
pid==3.0.4
promise==2.3
psycopg2-binary==2.9.9
pydantic==1.10.13
pydantic_core==2.10.1
pymongo==4.5.0
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
pyzmq==25.1.1
regex==2023.10.3
Rx==1.6.3
service==0.6.0
setproctitle==1.3.3
singledispatch==3.7.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.28
sqlalchemy2-stubs==0.0.2a35
sqlmodel==0.0.16
starlette==0.27.0
strawberry-graphql==0.220.0
tornado==6.3.3
typing_extensions==4.8.0
tzlocal==5.1
uvicorn==0.23.2
Werkzeug==3.0.0
